krikos_relay

Module defaults

Default values used in the relay.

Modules

timeoutsNon-wasm_browser

Contains all timeouts that we use in krikos.

Constants

DEFAULT_HTTPS_PORT

The default HTTPS port used by the Relay server.

DEFAULT_HTTP_PORT

The default HTTP port used by the Relay server.

DEFAULT_KEY_CACHE_CAPACITY

The default capacity of the key cache for the relay server.

DEFAULT_METRICS_PORT

The default metrics port used by the Relay server.

DEFAULT_RELAY_QUIC_PORT

The default QUIC port used by the Relay server to accept QUIC connections for QUIC address discovery