krikos_relay::protos

Module relay

This module implements the send/recv relaying protocol.

Protocol flow:

server occasionally sends FrameType::Ping

client responds to any FrameType::Ping with a FrameType::Pong

clients sends FrameType::ClientToRelayDatagram or FrameType::ClientToRelayDatagramBatch

server then sends FrameType::RelayToClientDatagram or FrameType::RelayToClientDatagramBatch to recipient

server sends FrameType::EndpointGone when the other client disconnects

Structs

Datagrams

One or multiple datagrams being transferred via the relay.

Enums

ClientToRelayMsg

Messages that clients send to relays.

Error

Protocol send errors.

RelayToClientMsg

The messages that a relay sends to clients or the clients receive from the relay.

Status

One-way message from server to client indicating issues with the relay connection.

Constants

MAX_PACKET_SIZE

The maximum size of a packet sent over relay. (This only includes the data bytes visible to the socket, not including its on-wire framing overhead)

MAX_RESTART_DURATION_MILLIS

Largest restart duration representable by the relay wire protocol, in milliseconds.

PER_CLIENT_SEND_QUEUE_DEPTHserver

The number of packets buffered for sending per client