krikos_relay::server::testing

Function tls_config

pub fn tls_config() -> TlsConfig

Available on with_crypto_provider and crate feature server and crate feature test-utils only.

Creates a TlsConfig suitable for testing.

Uses a self signed certificate valid for the "localhost" and "127.0.0.1" domains.

Configures https to be served on an OS assigned port on ipv4.